+++ src/Phan/Plugin/ConfigPluginSet.php
@@ -51,6 +51,7 @@ use Phan\PluginV2\AnalyzeMethodCapability;
 use Phan\PluginV2\AnalyzePropertyCapability;
 use Phan\PluginV2\AutomaticFixCapability;
 use Phan\PluginV2\BeforeAnalyzeCapability;
+use Phan\PluginV2\BeforeAnalyzePhaseCapability;
 use Phan\PluginV2\BeforeAnalyzeFileCapability;
 use Phan\PluginV2\FinalizeProcessCapability;
 use Phan\PluginV2\HandleLazyLoadInternalFunctionCapability;
@@ -89,6 +90,7 @@ final class ConfigPluginSet extends PluginV2 implements
     AnalyzeMethodCapability,
     AnalyzePropertyCapability,
     BeforeAnalyzeCapability,
+    BeforeAnalyzePhaseCapability,
     BeforeAnalyzeFileCapability,
     FinalizeProcessCapability,
     ReturnTypeOverrideCapability,
@@ -116,9 +118,14 @@ final class ConfigPluginSet extends PluginV2 implements
     private $before_analyze_file_plugin_set;
 
     /**
-     * @var array<int,BeforeAnalyzeCapability> - plugins to analyze the project before Phan starts the analyze phase.
+     * @var array<int,BeforeAnalyzeCapability> - plugins to analyze the project before Phan starts the analyze phase and before methods are analyzed.
      */
     private $before_analyze_plugin_set;
+
+    /**
+     * @var array<int,BeforeAnalyzePhaseCapability> - plugins to analyze the project before Phan starts the analyze phase and after methods are analyzed.
+     */
+    private $before_analyze_phase_plugin_set;
 
     /**
      * @var array<int,AfterAnalyzeFileCapability> - plugins to analyze files after Phan's analysis of that file is completed.
@@ -316,6 +323,8 @@ final class ConfigPluginSet extends PluginV2 implements
     }
 
     /**
+     * This method is called before analyzing a project and before analyzing methods.
+     *
      * @param CodeBase $code_base
      * The code base in which the project exists
      *
@@ -326,6 +335,22 @@ final class ConfigPluginSet extends PluginV2 implements
         $this->did_analyze_phase_start = true;
         foreach ($this->before_analyze_plugin_set as $plugin) {
             $plugin->beforeAnalyze($code_base);
+        }
+    }
+
+    /**
+     * This method is called before analyzing a project and after analyzing methods.
+     *
+     * @param CodeBase $code_base
+     * The code base in which the project exists
+     *
+     * @override
+     */
+    public function beforeAnalyzePhase(CodeBase $code_base)
+    {
+        $this->did_analyze_phase_start = true;
+        foreach ($this->before_analyze_phase_plugin_set as $plugin) {
+            $plugin->beforeAnalyzePhase($code_base);
         }
     }
 
@@ -850,6 +875,7 @@ final class ConfigPluginSet extends PluginV2 implements
         $this->pre_analyze_node_plugin_set      = self::filterPreAnalysisPlugins($plugin_set);
         $this->post_analyze_node_plugin_set     = self::filterPostAnalysisPlugins($plugin_set);
         $this->before_analyze_plugin_set        = self::filterByClass($plugin_set, BeforeAnalyzeCapability::class);
+        $this->before_analyze_phase_plugin_set  = self::filterByClass($plugin_set, BeforeAnalyzePhaseCapability::class);
         $this->before_analyze_file_plugin_set   = self::filterByClass($plugin_set, BeforeAnalyzeFileCapability::class);
         $this->after_analyze_file_plugin_set    = self::filterByClass($plugin_set, AfterAnalyzeFileCapability::class);
         $this->analyze_method_plugin_set        = self::filterByClass($plugin_set, AnalyzeMethodCapability::class);